Fixed the camera
Camera now functions 100% as intended - when you right click, it'll
somewhat throw the camera when you release. In other words, when you let
go of right click while still moving the camera, it won't just dead
stop.. instead it'll gently glide to a halt but at a reasonable pace so
as to not be silly.