Added save base classes
Need to figure out how the heck we want to go about handling it.  Most
of the pain is going to be loading it back rather than saving.

I'm thinking we cycle through all our structures and enemies and save
their name, location, rotation, scale, and then other attributes like
health, power, etc..  Then when we load, we go through and replace each
one based on those stats and then the hardest part is going to be
rebuilding the power tree.

When we first put them back in, none of them will have any nodes in the
power tree and therefore simply placing them all and then calling
RebuildTrees() probably won't work.

We could maybe call the structure placement function for each structure
as we place it back down that way they all get added as we go.  We can
start with sources and go up from there.  Unfortunately, that means we
may end up getting beams in different positions - for example if a
structure already was at the beam limit and had a couple extra
structures around it, depending on the order they're replaced, the
layout might be different and then that means some structures may even
get separated from the tree as a result.